Render to  screen

+++ src/filters/blur_filter.rs
@@ -243,7 +243,7 @@ impl BlurFilter {
                 }],
                 // ??????
                 depth_stencil_state: None,
-                index_format: wgpu::IndexFormat::Uint16,
+                index_format: wgpu::IndexFormat::Uint32,
                 vertex_buffers: &[wgpu::VertexBufferDescriptor {
                     stride: vertex_size as wgpu::BufferAddress,
                     step_mode: wgpu::InputStepMode::Vertex,
